requirements.txt
torch
transformers
Pillow
gradio
dotenv
huggingface_hub
requests